pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ledgermind"
version = "0.1.0"
description = "Privacy-first, read-only YNAB financial planning assistant (CLI + MCP)"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [{ name = "Yeixio" }]
dependencies = [
    "httpx>=0.27,<1",
    "mcp>=1.0,<2",
    "pydantic>=2.6,<3",
    "pydantic-settings>=2.2,<3",
    "pyyaml>=6,<7",
    "typer>=0.12,<1",
    "rich>=13,<15",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "respx>=0.21,<1",
    "ruff>=0.6,<1",
    "mypy>=1.11,<2",
    "types-PyYAML>=6,<7",
    "pre-commit>=3.7,<4",
    "mkdocs>=1.6,<2",
]
api = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
]
all = ["ledgermind[dev,api]"]

[project.scripts]
ledgermind = "ledgermind.cli:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
mypy_path = "src"
packages = ["ledgermind"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
